// contracts/generated/typescript/models/RollbackChangeBatchResult.ts
export interface RollbackChangeBatchResult {
    /**
     *
     * @type {string}
     * @memberof RollbackChangeBatchResult
     */
    changeBatchId: string;
    /**
     *
     * @type {Array<string>}
     * @memberof RollbackChangeBatchResult
     */
    revisionIds: Array<string>;
    /**
     *
     * @type {Array<string>}
     * @memberof RollbackChangeBatchResult
     */
    compensationClaimIds: Array<string>;
    /**
     *
     * @type {Array<string>}
     * @memberof RollbackChangeBatchResult
     */
    compensatedPageIds: Array<string>;
    /**
     *
     * @type {Array<string>}
     * @memberof RollbackChangeBatchResult
     */
    deletedPageIds: Array<string>;
    /**
     *
     * @type {Array<string>}
     * @memberof RollbackChangeBatchResult
     */
    deletedEntityIds: Array<string>;
    /**
     *
     * @type {Array<string>}
     * @memberof RollbackChangeBatchResult
     */
    restoredEntityIds: Array<string>;
    /**
     *
     * @type {Array<string>}
     * @memberof RollbackChangeBatchResult
     */
    entityMergeIds: Array<string>;
    /**
     *
     * @type {Array<string>}
     * @memberof RollbackChangeBatchResult
     */
    collectionIds: Array<string>;
    /**
     *
     * @type {number}
     * @memberof RollbackChangeBatchResult
     */
    removedClaimSourceCount: number;
    /**
     *
     * @type {boolean}
     * @memberof RollbackChangeBatchResult
     */
    idempotent: boolean;
}